refactor(table-container): extract types & interfaces into a separate file
