+++ src/app/components/observers/observers.component.html
@@ -1,4 +1,4 @@
-<div class="container-fluid">
+<div class="container-fluid c-observers__container">
   
   <div class="d-flex justify-content-between mt-2 mb-4 pl-0 pr-2 c-observers__header">
     <div class="d-flex align-items-center">
