+++ src/app/components/observers/observers.component.ts
@@ -27,10 +27,10 @@ import { ToastrService } from 'ngx-toastr';
 import { NgbModalRef, NgbModal, NgbModalOptions } from '@ng-bootstrap/ng-bootstrap';
 import { HttpClient } from '@angular/common/http';
 import { BASE_BUTTON_VARIANTS, Variants } from 'src/app/shared/base-button/base-button.component';
-import { SelectedZoneEvents, SortedColumnEvent, TableColumn, TableColumnTranslated } from 'src/app/table/table-container/table-container.component';
 import { DropdownConfigItem } from 'src/app/shared/base-dropdown/base-dropdown.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
+import { TableColumn, TableColumnTranslated, SelectedZoneEvents, SortedColumnEvent } from 'src/app/table/table.model';
 
 const ACTIONS_COLUMN_NAME = 'Actions';
