+++ src/app/table/table-container/table-container.component.ts
@@ -1,39 +1,7 @@
 import { ChangeDetectionStrategy, Component, ContentChild, ContentChildren, EventEmitter, HostBinding, Inject, InjectionToken, Input, OnInit, Output, TemplateRef } from '@angular/core';
-import { Observable } from 'rxjs';
 import { BASE_BUTTON_VARIANTS, Variants } from 'src/app/shared/base-button/base-button.component';
 
-export interface TableColumn {
-  name: string;
-  canBeSorted?: boolean;
-  propertyName?: string;
-}
-
-export type TableColumnTranslated = Omit<TableColumn, 'name'> & { name: Observable<any> }
-
-export enum SortDirection {
-  ASC,
-  DESC
-}
-
-export interface SortedColumnEvent {
-  col: TableColumn;
-  sortDirection: SortDirection;
-}
-
-export enum SelectedZoneEvents {
-  DELETE,
-  NOTIFCATION
-}
-
-const SELECTED_ZONE_EVENTS = new InjectionToken('SELECTED_ZONE_EVENTS', {
-  providedIn: 'root',
-  factory: () => SelectedZoneEvents,
-});
-
-const SORT_DIRECTION = new InjectionToken('SORT_DIRECTION', {
-  providedIn: 'root',
-  factory: () => SortDirection,
-});
+import { SelectedZoneEvents, SortDirection, SortedColumnEvent, TableColumn, SELECTED_ZONE_EVENTS, SORT_DIRECTION } from '../table.model';
 
 @Component({
   selector: 'app-table-container[columns]',
